move "create"-related storage code into a separate file.
